backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
python-dotenv==1.0.1
email-validator==2.2.0
